Refactor class naming and references for consistency in transaction processing and service layers
