+++ src/test/java/info/mackiewicz/bankapp/system/transaction/processing/core/execution/impl/OwnTransferExecutorTest.java
@@ -7,22 +7,20 @@ import info.mackiewicz.bankapp.transaction.model.TransactionType;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.InOrder;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 import java.math.BigDecimal;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyInt;
-import static org.mockito.Mockito.*;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
 
 /**
- * Tests for the FeeTransactionCommand class.
+ * Tests for the OwnTransferExecutor class.
  */
 @ExtendWith(MockitoExtension.class)
-class FeeTransactionCommandTest {
+class OwnTransferExecutorTest {
 
     @Mock
     private AccountService accountService;
@@ -34,22 +32,19 @@ class FeeTransactionCommandTest {
     private Account destinationAccount;
     
     @Mock
-    private Account bankAccount;
-    
-    @Mock
     private Transaction transaction;
 
-    private FeeTransactionExecutor command;
+    private OwnTransferExecutor command;
     private BigDecimal amount;
 
     @BeforeEach
     void setUp() {
-        command = new FeeTransactionExecutor();
-        amount = new BigDecimal("10.00");
+        command = new OwnTransferExecutor();
+        amount = new BigDecimal("100.00");
     }
 
     @Test
-    void execute_WithDestinationAccount_ShouldWithdrawAndDeposit() {
+    void execute_ShouldWithdrawFromSourceAndDepositToDestination() {
         // Arrange
         when(transaction.getSourceAccount()).thenReturn(sourceAccount);
         when(transaction.getDestinationAccount()).thenReturn(destinationAccount);
@@ -61,35 +56,14 @@ class FeeTransactionCommandTest {
         // Assert
         verify(accountService).withdraw(sourceAccount, amount);
         verify(accountService).deposit(destinationAccount, amount);
-        verify(transaction, never()).setDestinationAccount(any());
-        verify(accountService, never()).getAccountById(anyInt());
     }
 
     @Test
-    void execute_WithoutDestinationAccount_ShouldUseAndSetBankAccount() {
-        // Arrange
-        when(transaction.getSourceAccount()).thenReturn(sourceAccount);
-        when(transaction.getDestinationAccount()).thenReturn(null);
-        when(transaction.getAmount()).thenReturn(amount);
-        when(accountService.getAccountById(-1)).thenReturn(bankAccount);
-        
-        // Act
-        command.execute(transaction, accountService);
-        
-        // Assert
-        InOrder inOrder = inOrder(accountService, transaction);
-        inOrder.verify(accountService).getAccountById(-1);
-        inOrder.verify(transaction).setDestinationAccount(bankAccount);
-        inOrder.verify(accountService).withdraw(sourceAccount, amount);
-        inOrder.verify(accountService).deposit(bankAccount, amount);
-    }
-
-    @Test
-    void getTransactionType_ShouldReturnFee() {
+    void getTransactionType_ShouldReturnTransferOwn() {
         // Act
         TransactionType result = command.getTransactionType();
         
         // Assert
-        assertEquals(TransactionType.FEE, result);
+        assertEquals(TransactionType.TRANSFER_OWN, result);
     }
 }
